Some Refactoring

+++ Convertor/slice.h
@@ -5,7 +5,7 @@
 
 size_t slice_length(slice slice)
 {
-	double fractional = (double)(slice.high - slice.low) / slice.step;
+	double fractional = (double)(slice.high - slice.low) / (double)slice.step;
 	size_t plus_one = (fractional - (size_t)(fractional)) == 0 ? 0 : 1;
 
 	return (size_t)fractional + plus_one;
